requirements.txt
# Video Transcriber - Core Dependencies
# Minimal installation with essential features for video transcription
# For complete AI features, use: pip install -r requirements-full.txt

bcrypt>=4.0.0              # Password hashing

# Background task processing
celery>=5.3.0              # Background task queue
email-validator>=2.0.0     # Email validation

# Media processing and system utilities
ffmpeg-python>=0.2.0

# Web framework and real-time communication
flask>=3.0.0

# Authentication and Security (required by application)
Flask-Login>=0.6.3         # User authentication
Flask-Migrate>=4.0.0       # Database migrations
flask-socketio>=5.5.0
Flask-SQLAlchemy>=3.0.0    # Database ORM
Flask-WTF>=1.2.0           # Form handling and CSRF protection
numpy>=2.0.0
# Core transcription engine
openai-whisper>=20231117
psutil>=7.0.0
redis>=5.0.0               # Message broker for Celery
requests>=2.25.0          # HTTP client for health checks

# Mathematical and ML foundation
torch>=2.0.0
WTForms>=3.0.0             # Form validation

# ========================================
# UPGRADE TO FULL FEATURES
# ========================================
#
# For AI insights and advanced features, install:
# pip install -r requirements-full.txt
#
# Or run the upgrade script:
# python scripts/install_ai_features.py
#
# Full features include:
# - Sentiment analysis (TextBlob)
# - Topic modeling (Scikit-learn)
# - Advanced NLP (SpaCy)
# - PDF reports (ReportLab)
# - DOCX exports (python-docx)
